testsite/requirements.txt
aiohttp==3.5.4
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-extensions==2.2.1
django-rest-auth==0.9.5
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
gevent==1.4.0
greenlet==0.4.15
grequests==0.4.0
h11==0.8.1
h2==3.1.1
hpack==3.0.0
hstspreload==2019.8.20
hyperframe==5.2.0
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
Pillow==6.1.0
pipdate==0.3.2
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
rfc3986==1.3.2
six==1.12.0
sqlparse==0.3.0
typing-extensions==3.7.4
urllib3==1.25.3
websockets==8.0.2
yarl==1.3.0
